last update migrations

+++ database/migrations/2024_09_20_233115_create_email_sents_table.php
@@ -8,16 +8,18 @@ class CreateEmailSentsTable extends Migration
 {
     public function up()
     {
-        Schema::create('email_sents', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('recap_id');
-            $table->string('recipient_email');
-            $table->timestamp('sent_at')->nullable();
-            $table->enum('status', ['sent', 'failed']);
-            $table->timestamps();
+        if (!Schema::hasTable('email_sents')) {
+            Schema::create('email_sents', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('recap_id');
+                $table->string('recipient_email');
+                $table->timestamp('sent_at')->nullable();
+                $table->enum('status', ['sent', 'failed']);
+                $table->timestamps();
 
-            $table->foreign('recap_id')->references('id')->on('recaps')->onDelete('cascade');
-        });
+                $table->foreign('recap_id')->references('id')->on('recaps')->onDelete('cascade');
+            });
+        }
     }
 
     public function down()
